refactor webpack config
