+++ config/webpack/webpack.common.js
@@ -1,22 +1,12 @@
-const env = require('dotenv').config({ silent: true });
 const webpack = require('webpack');
-const path = require('path');
-const _ = require('lodash');
+const webpackMerge = require('webpack-merge');
 const HtmlWebpackPlugin = require('html-webpack-plugin');
 const ExtractTextPlugin = require('extract-text-webpack-plugin');
 
 const helpers = require('./helpers');
-const constants = require('./constants');
+const coreConfig = require('./webpack.core');
 
 const isProd = process.env.npm_lifecycle_event === 'build';
-const envMap = _.mapValues(env, v => JSON.stringify(v));
-
-if (!envMap.APP_ENV) {
-  envMap.APP_ENV = 'development';
-  console.log('APP_ENV is not set in your .env, it will default to "development"');
-} else {
-  console.log('APP_ENV is ' + envMap.APP_ENV);
-}
 
 const entry = {
   'polyfills': './src/polyfills.ts',
@@ -25,13 +15,8 @@ const entry = {
   'app': './src/main.ts'
 };
 
-module.exports = {
+module.exports = webpackMerge(coreConfig, {
   entry: entry,
-
-  resolve: {
-    extensions: ['.js', '.ts'],
-    modules: ['node_modules', helpers.root('src')]
-  },
 
   module: {
     rules: [
@@ -71,21 +56,6 @@ module.exports = {
 
   plugins: [
     new webpack.NamedModulesPlugin(),
-    new webpack.ProvidePlugin({
-        jQuery: 'jquery',
-        $: 'jquery',
-        jquery: 'jquery'
-    }),
-
-    new webpack.DefinePlugin({
-      'process.env': envMap
-    }),
-
-    new webpack.ContextReplacementPlugin(
-      // The (\\|\/) piece accounts for path separators in *nix and Windows
-      constants.CONTEXT_REPLACE_REGEX,
-      helpers.root('./src') // location of your src
-    ),
 
     new webpack.optimize.CommonsChunkPlugin({
       name: ['app', 'vendor', 'vendorDll', 'style', 'polyfills']
@@ -107,4 +77,4 @@ module.exports = {
     })
   ]
 
-};
+});
